About page is now the homepage instead of the blog.
Blog now navigable but not as prominent. Move tools into its own page.

Tags-based listings are proper sub-blogs in their own right (instead of
just lists of articles) and will probably be the better way to navigate
the blog in the future. Link to a couple of those tags but a big
reclassificaiton pass is needed.